The dimension of the matrizes was incremented. The number of rows and columns are long. Several changes into the methods to make compatible the variables with new dimensions.

+++ LinearSolvers/bicgstab.h
@@ -25,7 +25,7 @@
 template < class Matrix, class Vector, class Preconditioner, class Real >
 int 
 BiCGSTAB(const Matrix &A, Vector &x, const Vector &b,
-         /*const */Preconditioner &M, int &max_iter, Real &tol, Vector *residual,const int FromCurrent)
+         /*const */Preconditioner &M, long &max_iter, Real &tol, Vector *residual,const int FromCurrent)
 {
 	Real resid;
 	Vector rho_1(1), rho_2(1), alpha(1), beta(1), omega(1);
@@ -54,7 +54,7 @@ BiCGSTAB(const Matrix &A, Vector &x, const Vector &b,
 		return 0;
 	}
 	
-	for (int i = 1; i <= max_iter; i++) {
+	for (long i = 1; i <= max_iter; i++) {
 		rho_1(0) = Dot(rtilde, r);
 		if (rho_1(0) == 0) {
 			tol = Norm(r) / normb;
