+++ Material/REAL/pzviscoelastic.cpp
@@ -250,7 +250,7 @@ void TPZViscoelastic::Solution(TPZMaterialData &data, int var, TPZVec<STATE> &So
 		TPZFNMatrix<9,STATE> StrainTensor(3,3);
 		TPZFMatrix<STATE> DSol(data.dsol[0]);
 		TPZMatWithMem<TPZFMatrix<STATE>,TPZElasticity3D>::ComputeStrainTensor(StrainTensor, DSol);
-		int numiterations = 1000;
+		long numiterations = 1000;
 		REAL tol = TPZElasticity3D::gTolerance;
         TPZManVector<STATE,3> eigv(3,0.);
 		bool result;
