+++ Matrix/pzmatred.cpp
@@ -34,7 +34,7 @@ TPZMatRed<TVar,  TSideMatrix>::TPZMatRed () : TPZMatrix<TVar>( 0, 0 ), fK11(0,0)
 }
 
 template<class TVar, class TSideMatrix>
-TPZMatRed<TVar, TSideMatrix>::TPZMatRed( int dim, int dim00 ):TPZMatrix<TVar>( dim,dim ), fK11(dim-dim00,dim-dim00,0.), fK01(dim00,dim-dim00,0.), 
+TPZMatRed<TVar, TSideMatrix>::TPZMatRed( long dim, long dim00 ):TPZMatrix<TVar>( dim,dim ), fK11(dim-dim00,dim-dim00,0.), fK01(dim00,dim-dim00,0.),
 fK10(dim-dim00,dim00,0.), fF0(dim00,1,0.),fF1(dim-dim00,1,0.), fMaxRigidBodyModes(0), fNumberRigidBodyModes(0)
 {
 	if(dim<dim00) TPZMatrix<TVar>::Error(__PRETTY_FUNCTION__,"dim k00> dim");
@@ -65,7 +65,7 @@ void TPZMatRed<TVar, TSideMatrix>::SimetrizeMatRed() {
 	fK11.Simetrize();
 	
 	/*
-	int row,col;
+	long row,col;
 	for(row=0; row<fDim1; row++) {
 		for(col=row+1; col<fDim1; col++) {
 			(fK11)(col,row) = (fK11)(row,col);
@@ -76,8 +76,8 @@ void TPZMatRed<TVar, TSideMatrix>::SimetrizeMatRed() {
 
 template<class TVar, class TSideMatrix>
 int
-TPZMatRed<TVar, TSideMatrix>::PutVal(const int r,const int c,const TVar& value ){
-	int row(r),col(c);
+TPZMatRed<TVar, TSideMatrix>::PutVal(const long r,const long c,const TVar& value ){
+	long row(r),col(c);
 	if (IsSimetric() && row > col ) Swap( &row, &col );
 	if (row<fDim0 &&  col<fDim0)  fK00->PutVal(row,col,value);
 	if (row<fDim0 &&  col>=fDim0)  fK01.PutVal(row,col-fDim0,(TVar)value);
@@ -89,8 +89,8 @@ TPZMatRed<TVar, TSideMatrix>::PutVal(const int r,const int c,const TVar& value )
 
 template<class TVar, class TSideMatrix>
 const TVar&
-TPZMatRed<TVar,TSideMatrix>::GetVal(const int r,const int c ) const {
-	int row(r),col(c);
+TPZMatRed<TVar,TSideMatrix>::GetVal(const long r,const long c ) const {
+	long row(r),col(c);
 	
 	if (IsSimetric() && row > col ) Swap( &row, &col );
 	if (row<fDim0 &&  col<fDim0)  return ( fK00->GetVal(row,col) );
@@ -101,8 +101,8 @@ TPZMatRed<TVar,TSideMatrix>::GetVal(const int r,const int c ) const {
 }
 
 template<class TVar, class TSideMatrix>
-TVar& TPZMatRed<TVar,TSideMatrix>::s(const int r,const int c ) {
-	int row(r),col(c);
+TVar& TPZMatRed<TVar,TSideMatrix>::s(const long r,const long c ) {
+	long row(r),col(c);
 	
 	if (r < fDim0 && IsSimetric() && row > col ) Swap( &row, &col );
 	if (row<fDim0 &&  col<fDim0)  return ( fK00->s(row,col) );
@@ -131,7 +131,7 @@ template<class TVar, class TSideMatrix>
 void TPZMatRed<TVar,TSideMatrix>::SetF(const TPZFMatrix<TVar> & F)
 {
 	
-	int FCols=F.Cols(),c,r,r1;
+	long FCols=F.Cols(),c,r,r1;
 	
 	fF0.Redim(fDim0,FCols);
 	fF1.Redim(fDim1,FCols);
@@ -290,7 +290,7 @@ void TPZMatRed<REAL, TPZVerySparseMatrix<REAL> >::UGlobal(const TPZFMatrix<REAL>
 	fK01.MultAdd(U1,(F0),u0,-1,0);
 	
 	result.Redim( fDim0+fDim1,F0.Cols() );
-	int c,r,r1;
+	long c,r,r1;
 	
 	for(c=0; c<F0.Cols(); c++)
 	{
@@ -340,7 +340,7 @@ void TPZMatRed<TVar, TSideMatrix >::UGlobal(const TPZFMatrix<TVar> & U1, TPZFMat
 #endif
 	
 	result.Redim( fDim0+fDim1,fF0.Cols() );
-	int c,r,r1;
+	long c,r,r1;
 	
 	for(c=0; c<fF0.Cols(); c++)
 	{
@@ -390,9 +390,9 @@ void TPZMatRed<TVar, TSideMatrix>::UGlobal2(TPZFMatrix<TVar> & U1, TPZFMatrix<TV
 #endif
 	
 	result.Redim( fDim0+fDim1,fF0.Cols() );
-    int nglob = fDim0+fDim1;
+    long nglob = fDim0+fDim1;
 	
-	int c,r,r1;
+	long c,r,r1;
 	
 	for(c=0; c<fF0.Cols(); c++)
 	{
@@ -433,7 +433,7 @@ void TPZMatRed<TVar, TSideMatrix>::Print(const char *name , std::ostream &out ,c
 }
 
 template<class TVar, class TSideMatrix>
-int TPZMatRed<TVar,TSideMatrix>::Redim(int dim, int dim00){
+int TPZMatRed<TVar,TSideMatrix>::Redim(long dim, long dim00){
 	if(dim<dim00) TPZMatrix<TVar>::Error(__PRETTY_FUNCTION__,"dim k00> dim");
 	if(fK00) fK00->Redim(dim00,dim00);
 	if(fF0)  delete fF0;
