+++ Matrix/pzmatred.h
@@ -49,7 +49,7 @@ public:
 	 * @param dim assumes the value of n1+n2
 	 * @param dim00 equals n1
 	 */
-	TPZMatRed(const int dim, const int dim00);
+	TPZMatRed(const long dim, const long dim00);
 	
 	template<class TSideCopy>
 	TPZMatRed<TVar ,TSideMatrix>(const TPZMatRed<TVar, TSideCopy> &cp): TPZMatrix<TVar>(cp), fK11(cp.fK11), fK01(cp.fK01), fK10(cp.fK10), fF0(cp.fF0), fF1(cp.fF1),fMaxRigidBodyModes(cp.fMaxRigidBodyModes),fNumberRigidBodyModes(cp.fNumberRigidBodyModes)
@@ -81,9 +81,9 @@ public:
 	 * @brief Put and Get values without bounds checking
 	 * these methods are faster than "Put" e "Get" if DEBUG is defined
 	 */
-	virtual int PutVal(const int row, const int col, const TVar& value);
-	virtual const TVar &GetVal(const int row, const int col) const;
-	virtual TVar &s(const int row, const int col);
+	virtual int PutVal(const long row, const long col, const TVar& value);
+	virtual const TVar &GetVal(const long row, const long col) const;
+	virtual TVar &s(const long row, const long col);
 	
 	/** @brief This method will zero all submatrices associated with this reducable matrix class */
 	virtual int Zero();
@@ -107,12 +107,12 @@ public:
 		return fK10;
 	}
     
-    int Dim0()
+    long Dim0()
     {
         return fDim0;
     }
     
-    int Dim1()
+    long Dim1()
     {
         return fDim1;
     }
@@ -166,7 +166,7 @@ public:
 			   const MatrixOutputFormat = EFormatted) const;
 	
 	/** @brief Redim: Set the dimension of the complete matrix and reduced matrix */
-	int Redim(const int dim,const int dim00); //Cesar 19/12/00
+	int Redim(const long dim,const long dim00); //Cesar 19/12/00
 	
 	/**
 	 * @brief It computes z = beta * y + alpha * opt(this)*x but z and x can not overlap in memory.
@@ -199,7 +199,7 @@ private:
 	 * @param row Row number
 	 * @param col Column number
 	 */
-	static void Swap(int *row, int *col);
+	static void Swap(long *row, long *col);
     
     /** @brief Decompose K00 and adjust K01 and K10 to reflect rigid body modes */
     void DecomposeK00();
@@ -218,7 +218,7 @@ private:
 	TPZFMatrix<TVar> fF0, fF1;
 
 	/** @brief Stores matricess \f$ fKij \f$ dimensions */
-	int fDim0, fDim1;
+	long fDim0, fDim1;
 	
 	/** @brief Is true if the declared dimension of the matrix is fDim0 */
 	bool fIsReduced;
@@ -237,9 +237,9 @@ private:
 /*** Swap ***/
 /* @note Modificacao por Philippe Devloo (insercao de inline )*/
 template<class TVar, class TSideMatrix>
-inline void TPZMatRed<TVar, TSideMatrix>::Swap(int *a, int *b)
+inline void TPZMatRed<TVar, TSideMatrix>::Swap(long *a, long *b)
 {
-	int aux = *a;
+	long aux = *a;
 	*a = *b;
 	*b = aux;
 }
